Made the page reload to avoid bug
An unknown bug would cause the old views json to remain valid and when one created a new new it would fuse the selected views for all created views until the page was reloaded. The bug was avoided (not fixed) by reloading the page after every submit of custom views. This is not ideal since it creates high traffic on the api, which is called on every load of the page